requirements.txt
mistralai
faiss-cpu
pymupdf
streamlit
python-dotenv
tiktoken
numpy